Add Context.with_stage() method
This method can be used to change the stage currently configured on the
kitipy.Context. This is used by the _init_stage_group_wrapper() function
to ensure that not only the executor is changed by the kctx.stage
property is changed too (this wasn't the case and thus was buggy).