Add a tasks property to kitipy TaskGroup
Until now, the TaskGroup was adding tasks and other groups into commands
property, which is a dict of cmd name as key and click.Command as value.

However this structure isn't adapted to kitipy needs: as tasks can be
fliltered, there might be multiple tasks with the same name but
different filters. For instance, a stack group with a deploy task,
inside a stage group. In such case, the deploy task is probably quite
different for prod and dev stages, so it makes a perfect case for
filters. But due to the commands structure, adding two deploy functions
would actually erase the first one. Hence the need for a different
property, with a different structure layout.

The new tasks property is a list of kitipy.Task and kitipy.Groups.