Add confirm_and_apply() to utils
This new function will help for the common use-case where a first
command like a dry-run has to be run, then a confirm message asking the
user if they want to apply some dangerous changes is printed and finally
the given changes are applied.

This function can be used for linting or formatting tasks, when source
code is rewritten by some tools. But it can also be used to confirm some
dangerous actions like running migrations on prod stage.

Also, this function is fully configurable: each step can be disabled
separately (dry run/ask for confirm/apply) and each step is a callable.

For now only bool-based confirmation can be used (e.g. yes/no), but it's
planned to add string-based confirmation. With such confirmation, it
could be possible to ask to retype the name of the stage before doing
anything wrong, for instance.